+++ tests/features/share.php
@@ -15,15 +15,15 @@ class ShareTest extends PHPUnit_Framework_TestCase
             'foo' => 'Hello',
             'bar' => 'world',
         ]);
-        $html = $pug->render("p=foo\ndiv=answear");
+        $html = $pug->render("p=\$foo\ndiv=\$answear");
         $this->assertSame('<p>Hello</p><div>42</div>', $html);
 
-        $html = $pug->render("p=foo\ndiv=answear", [
+        $html = $pug->render("p=\$foo\ndiv=\$answear", [
             'answear' => 16,
         ]);
         $this->assertSame('<p>Hello</p><div>16</div>', $html);
 
-        $html = $pug->render("p\n  =foo\n  =' '\n  =bar\n  | !");
+        $html = $pug->render("p\n  ?=\$foo\n  =' '\n  =\$bar\n  | !");
         $this->assertSame('<p>Hello world!</p>', $html);
     }
 
@@ -41,11 +41,11 @@ class ShareTest extends PHPUnit_Framework_TestCase
 
         $error = null;
         try {
-            $pug->render("p\n  =foo\n=' '\n=bar\n  | !");
+            $pug->render("p\n  ?=\$foo\n=' '\n=\$bar\n  | !");
         } catch (\Exception $e) {
             $error = $e->getMessage();
         }
 
-        $this->assertSame('Undefined variable: foo', $error);
+        $this->assertRegExp('/Undefined variable: foo/', $error);
     }
 }